+++ src/Symfony/Component/Form/AbstractTypeExtension.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Component\Form;
 
+use Symfony\Component\Form\Exception\LogicException;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
@@ -44,5 +45,23 @@ abstract class AbstractTypeExtension implements FormTypeExtensionInterface
      */
     public function configureOptions(OptionsResolver $resolver)
     {
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @deprecated since Symfony 4.2, use getExtendedTypes() instead.
+     */
+    public function getExtendedType()
+    {
+        if (!method_exists($this, 'getExtendedTypes')) {
+            throw new LogicException(sprintf('You need to implement the static getExtendedTypes() method when implementing the %s in %s.', FormTypeExtensionInterface::class, static::class));
+        }
+
+        @trigger_error(sprintf('The %s::getExtendedType() method is deprecated since Symfony 4.2 and will be removed in 5.0. Use getExtendedTypes() instead.', \get_class($this)), E_USER_DEPRECATED);
+
+        foreach (static::getExtendedTypes() as $extendedType) {
+            return $extendedType;
+        }
     }
 }