+++ src/Symfony/Component/Form/Tests/Fixtures/FooTypeBarExtension.php
@@ -28,8 +28,8 @@ class FooTypeBarExtension extends AbstractTypeExtension
         );
     }
 
-    public function getExtendedType()
+    public static function getExtendedTypes(): iterable
     {
-        return __NAMESPACE__.'\FooType';
+        return array(__NAMESPACE__.'\FooType');
     }
 }